Remove all but remaining upload actions from bundle manager.
