--- enterprise/cmd/precise-code-intel-bundle-manager/internal/paths/paths_test.go
@@ -1,17 +1,0 @@
-package paths
-
-import (
-	"flag"
-	"os"
-	"testing"
-
-	"github.com/inconshreveable/log15"
-)
-
-func TestMain(m *testing.M) {
-	flag.Parse()
-	if !testing.Verbose() {
-		log15.Root().SetHandler(log15.DiscardHandler())
-	}
-	os.Exit(m.Run())
-}
